- Updated the Trakt account empty state design to fit well with larger screens - Updated the favorites list composable to adapt to the different window size classes - Fixed the ListPosterCard image width which was not scaling well across different windowsize classes
