+++ app/src/main/java/com/theupnextapp/extensions/ReferenceDevices.kt
@@ -15,8 +15,8 @@ package com.theupnextapp.extensions
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
 
-@Preview(name = "phone", device = Devices.PHONE)
-@Preview(name = "foldable", device = Devices.FOLDABLE)
-@Preview(name = "custom", device = "spec:width=1280dp,height=800dp,dpi=480")
-@Preview("desktop", device = "id:desktop_medium")
+@Preview(name = "phone", device = Devices.PHONE, showBackground = true)
+@Preview(name = "foldable", device = Devices.FOLDABLE, showBackground = true)
+@Preview(name = "custom", device = "spec:width=1280dp,height=800dp,dpi=480", showBackground = true)
+@Preview("desktop", device = "id:desktop_medium", showBackground = true)
 annotation class ReferenceDevices
